Add numeric representation translation

+++ src/main.rs
@@ -195,7 +195,7 @@ fn c_compile_source(input: &str) -> Result<String, Error> {
          \n\
          \ntypedef union _Aro_Any {{\
          \n  bool Bool;\
-         \n  double Int;\
+         \n  int Int;\
          \n  double Float;\
          \n  void* Void_Ptr;\
          \n  union _Aro_Any* Any_Ptr;\
